Delete encoders.
