--- src/tabpfn/architectures/encoders/steps/__init__.py
@@ -1,27 +1,0 @@
-#  Copyright (c) Prior Labs GmbH 2026.
-
-from .feature_group_projections_encoder_step import (
-    LinearInputEncoderStep,
-    MLPInputEncoderStep,
-)
-from .feature_transform_encoder_step import FeatureTransformEncoderStep
-from .frequency_feature_encoder_step import FrequencyFeatureEncoderStep
-from .multiclass_classification_target_encoder_step import (
-    MulticlassClassificationTargetEncoderStep,
-)
-from .nan_handling_encoder_step import NanHandlingEncoderStep
-from .normalize_feature_groups_encoder_step import NormalizeFeatureGroupsEncoderStep
-from .remove_duplicate_features_encoder_step import RemoveDuplicateFeaturesEncoderStep
-from .remove_empty_features_encoder_step import RemoveEmptyFeaturesEncoderStep
-
-__all__ = [
-    "FeatureTransformEncoderStep",
-    "FrequencyFeatureEncoderStep",
-    "LinearInputEncoderStep",
-    "MLPInputEncoderStep",
-    "MulticlassClassificationTargetEncoderStep",
-    "NanHandlingEncoderStep",
-    "NormalizeFeatureGroupsEncoderStep",
-    "RemoveDuplicateFeaturesEncoderStep",
-    "RemoveEmptyFeaturesEncoderStep",
-]
