--- src/tabpfn/model/transformer.py
@@ -1,7 +1,0 @@
-#  Copyright (c) Prior Labs GmbH 2026.
-
-"""DEPRECATED: Please import tabpfn.architectures.base.transformer instead."""
-
-from __future__ import annotations
-
-from tabpfn.architectures.base.transformer import *  # noqa: F403
